Added export and CSV program
Ryan took 4 commits to do this, 1 to actually write the code and 3 more to fix
bugs. He was also too lazy to squash the bug fixes together before merging into
master, so I took the liberty. For posterity, I included all the details below.
---------------

Addition of Export to CSV program

This code is 100% untested. I haven't even attempted to compile it yet. This is just a commit of work.
The program is complete, but not verified in any sense. Do not trust outputs! Feel free to bug test
I'll check everything later after work.

Found an error while reading through it.

Correction of the exporting program

This push is for testing with a larger database set.

Correction of the export program

The program now functions as intended